backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.12.0
python-multipart==0.0.6
diffusers==0.21.4
torch>=2.2.0
transformers
pillow
watchdog==3.0.0
pydantic==2.5.2
numpy==1.26.4
